01_backend/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
django-filter==23.5
django-environ==0.11.2
psycopg2-binary==2.9.9
channels==4.0.0
daphne==4.0.0
celery==5.3
django-celery-results==2.5.1
django-celery-beat==2.5.0
redis==4.5
channels-redis==4.1.0
paho-mqtt==1.6.1
boto3==1.34.1
django-storages==1.14.2
Pillow==10.1.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
django-debug-toolbar==4.3.0
pytest==7.4.4
pytest-django==4.5.2
black==23.12.1
isort==5.12.0
whitenoise==6.6.0
gunicorn==21.2.0
django-redis==5.4.0
django-extensions==3.2.3
django-cleanup==8.0.0
django-cors-headers==4.3.1
uvicorn==0.27.0